requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
httplib2==0.31.0
idna==3.11
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.9.0
pyparsing==3.3.1
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.2
